Add files via upload
This project started from a recent course I participated in on Udemy, ‘Microsoft SQL for Beginners’ by Brewster Knowlton. I took this course to refresh my skills with SQL. Throughout working, we used the AdventureWorks Database to practice queries, and once I finished it I decided that this database was a good opportunity to showcase the skills I learned for a project.